glm53-kpool(manager-merge,mechanical_provable): Consolidate DSA metadata management
Move KPool fusion selection and sibling replay methods into DSAMetadataManagementMixin in dsa_metadata_manager.py, as requested in review. Preserve the five method bodies and existing feature gates.
